Extract example service impl to their own files
